+++ src/compas/geometry/brep/face.py
@@ -41,6 +41,8 @@ class BrepFace(Data):
         Returns True if this face is a sphere, False otherwise.
     is_torus : bool, read-only
         Returns True if this face is a torus, False otherwise.
+    is_reversed : bool, read-only
+        True if the orientation of this face is reversed, False otherwise.
     is_valid : bool, read-only
         Return True if this face is valid, False otherwise.
     loops : list[:class:`compas.geometry.BrepLoop`], read-only
@@ -117,6 +119,10 @@ class BrepFace(Data):
 
     @property
     def centroid(self):
+        raise NotImplementedError
+
+    @property
+    def is_reversed(self):
         raise NotImplementedError
 
     @property